fix: move xcode helpers inside amplify-app
