+++ packages/amplify-app/src/xcodeHelpers.js
